feat: HTTP 상태 코드 상수 및 HTTPError 클래스 추가
