+++ packages/core/src/http/constants.ts
@@ -1,0 +1,9 @@
+export const HTTP_STATUS_CODE = {
+  SUCCESS: 200,
+  CREATED: 201,
+  BAD_REQUEST: 400,
+  UNAUTHORIZED: 401,
+  NOT_FOUND: 404,
+  CONFLICT: 409,
+  INTERNAL_SERVER_ERROR: 500,
+} as const;
