+++ packages/core/src/http/http-error.ts
@@ -1,25 +1,29 @@
-export const HTTP_STATUS_CODE = {
-  SUCCESS: 200,
-  CREATED: 201,
-  BAD_REQUEST: 400,
-  UNAUTHORIZED: 401,
-  NOT_FOUND: 404,
-  CONFLICT: 409,
-  INTERNAL_SERVER_ERROR: 500,
-} as const;
+import { HTTP_STATUS_CODE } from './constants';
 
-/** HTTP 통신 시 발생하는 API Error를 별도의 객체로 나타내기 위한 에러 객체입니다. */
+/**
+ * HTTP 통신 중 발생할 수 있는 에러를 나타내는 커스텀 에러 클래스입니다.
+ *
+ * 상태 코드에 따라 에러 이름(`name`)이 자동으로 설정되며,
+ * 선택적으로 서버에서 내려주는 도메인 에러 코드(`code`)도 포함할 수 있습니다.
+ */
 export class HTTPError extends Error {
   statusCode: number;
   code?: number;
 
+  /**
+   * HTTPError 인스턴스를 생성합니다.
+   *
+   * @param statusCode - HTTP 상태 코드
+   * @param message - 에러 메시지 (optional)
+   * @param code - 도메인 에러 코드 (optional)
+   */
   constructor(statusCode: number, message?: string, code?: number) {
     super(message);
     let name = 'HTTPError';
 
     switch (statusCode) {
       case HTTP_STATUS_CODE.BAD_REQUEST: {
-        name += ': BAD_REQUESET';
+        name += ': BAD_REQUEST';
         break;
       }
       case HTTP_STATUS_CODE.UNAUTHORIZED: {
@@ -30,9 +34,16 @@ export class HTTPError extends Error {
         name += ': NOT_FOUND';
         break;
       }
-
+      case HTTP_STATUS_CODE.CONFLICT: {
+        name += ': CONFLICT';
+        break;
+      }
       case HTTP_STATUS_CODE.INTERNAL_SERVER_ERROR: {
         name += ': INTERNAL_SERVER_ERROR';
+        break;
+      }
+      default: {
+        name += `: ${statusCode}`;
         break;
       }
     }
