+++ packages/core/src/http/index.ts
@@ -1,3 +1,4 @@
+export * from './constants';
 export * from './http-error';
 export * from './instance';
 export * from './types';
